controladores de horario y plan

+++ src/main/java/com/version/gymModuloControl/service/CategoriaService.java
@@ -1,4 +1,5 @@
 package com.version.gymModuloControl.service;
+import com.version.gymModuloControl.repository.ProductoRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import com.version.gymModuloControl.model.Categoria;
@@ -12,9 +13,14 @@ public class CategoriaService {
 
     @Autowired
     private CategoriaRepository categoriaRepository;
+
+    @Autowired
+    private ProductoRepository productoRepository;
     public List<Categoria> listarCategoria() {
         return categoriaRepository.findAll();
     }
+
+
 
     @Transactional
     public Categoria guardarCategoria(Categoria categoria) {
@@ -39,6 +45,10 @@ public class CategoriaService {
 
     @Transactional
     public boolean eliminarCategoria(Integer idCategoria) {
+        boolean existeEnProducto = productoRepository.existsByCategoria_IdCategoria(idCategoria);
+        if (existeEnProducto) {
+            throw new IllegalStateException("No se puede eliminar la categoría porque está asociada a un producto.");
+        }
         if (categoriaRepository.existsById(idCategoria)) {
             categoriaRepository.deleteById(idCategoria);
             return true;
